requirements.txt
django==3.2
djangorestframework==3.14.0
dj-rest-auth==5.0.1
django-allauth==0.58.0
django-sites
djangorestframework-simplejwt
django-cors-headers==4.1.0
python-dotenv
Pillow
